resume_agent/requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
Flask==2.3.3
Werkzeug==2.3.7
tabulate==0.9.0
 